New ComputationExpVal
With the missing operator error corrected I need to upload the functions to compute the expected values. Instead of one 1 I_if now we have 4 types and for each type (I^2, Ic^2, I*Ic, Is^2) i need to do the spin average over all the possible initial and final states.

Also i included a code that directly writes .txt files with the results of the I_if depnendece on the M