Code that we will also use to obtain A and B errors
After the crosschek we saw that linearizating the problem was wrong and hence the new way to obtain A, B and the spectrum's errors will be to consider the lattice data totally correlated. Therefore we find A+, B+ and A-, B- for lattice data +- lattice errors. Then the errors on A and B will be max(A+ minus A, A minus A-)